requirements.txt
anthropic>=0.25.0
flask>=3.0.0
pdfplumber>=0.10.0
sentence-transformers>=2.7.0
faiss-cpu>=1.8.0
numpy>=1.26.0
tqdm>=4.66.0
google-genai>=1.0.0
